Fix R11 pointer bug
FineIBT uses R11 to pass hashes across functions. This led to the
actual function pointer being overwritten when R11 is used by the
call instruction, causing a crash.

This patch fixes this bug by using the Register Scavenger API to
retrieve an available register to be used as auxiliar, then adding
instructions to move the pointer into the auxiliar before R11 is
overwritten with the hash and modifying the call to use it instead
of R11.

For obvious reasons this bug didnt manifest when the instrumented
binary was compiled with -ffixed-r11.